requirements.txt
fastapi
uvicorn
sqlmodel
pydantic
python-jose[cryptography]
python-dotenv
httpx
